backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
nltk>=3.8.1
rbo>=0.1.2
certifi>=2023.7.22
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
coverage>=7.2.5
mypy>=1.3.0
ruff>=0.0.262